Module dishub::ops::github

This module contains functions which interface with the GitHub API.

Every function in this module needs an instance of AppTokens, wherein the discord field does not need to be set.

The GitHub authentication is used just to get a bigger rate limit, so if you don't need to make a lot of requests just pass an empty string.

Functions

poll_repo_events_new
Get the events for a repository when you don't have an ETag (which is to say - for the first time).

poll_repo_events_update
Get the events for a repository when you already have an ETag (which is to say - after the first time).

poll_user_events_new
Get the events for a user when you don't have an ETag (which is to say - for the first time).

poll_user_events_update
Get the events for a user when you already have an ETag (which is to say - after the first time).

repo_exists
Check whether a repository with the specified slug exists.

user_exists
Check whether a user with the specified name exists.