Struct dishub::ops::Feed

pub struct Feed {
    pub subject: String,
    pub server: u64,
    pub channel: u64,
    pub e_tag: Option<String>,
    pub latest: Option<DateTime<FixedOffset>>,
    pub next_min: Option<DateTime<FixedOffset>>,
}

A feed to be checked on GitHub and sent to Discord.

Fields
subject: String

The thing to watch.

This can either be a user in the form: "username" or a repo slug in the form: "username/reponame".

server: u64

The Discord server ID to post to.

channel: u64

The Discord channel ID to post to.

e_tag: Option<String>

The ETag of the last received event pack.

latest: Option<DateTime<FixedOffset>>

The time of the last received event pack.

next_min: Option<DateTime<FixedOffset>>

The minimal time the next poll is allowed.

Methods

impl Feed

fn new(subject: String, server_id: u64, channel_id: u64) -> Feed

Create an unpolled feed with the specified subject and posting target.

fn read(p: &Path) -> Result<Vec<Feed>, Error>

Read the application feeds from the specified file.

fn write(feeds: Vec<Feed>, p: &Path)

Save the application feeds to the specified file.

fn poll(&mut self, tkn: &AppTokens) -> Result<Vec<Event>, Error>