Enum dishub::ops::EventPayload

pub enum EventPayload {
    CommitComment {
        content: String,
        commit_id: String,
        id: u64,
    },
    Create {
        ref_type: String,
        ref_name: Option<String>,
        master_branch: String,
        repo_description: String,
    },
    Delete {
        ref_type: String,
        ref_name: String,
    },
    Fork {
        new_slug: String,
    },
    Gollum {
        pages: Vec<GollumPayload>,
    },
    IssueComment {
        action: String,
        issue: u64,
        body: String,
        id: u64,
    },
    Issues {
        action: String,
        number: u64,
        title: String,
        body: String,
        labels: Vec<String>,
    },
    Member {
        action: String,
        user: String,
    },
    Public,
    PullRequest {
        action: String,
        number: u64,
        title: String,
        body: String,
        merged: bool,
    },
    PullRequestReview {
        action: String,
        pr: u64,
        state: String,
        body: String,
        id: u64,
    },
    PullRequestReviewComment {
        action: String,
        pr: u64,
        body: String,
        id: u64,
    },
    Push {
        pushed_ref: String,
        prev_head: String,
        new_head: String,
        size: u64,
        distinct_size: u64,
        commits: Vec<Commit>,
    },
    Release {
        action: String,
        tag_name: String,
        target: String,
        draft: bool,
        prerelease: bool,
        name: Option<String>,
        body: Option<String>,
    },
    Watch {
        action: String,
    },
    Other {
        event_type: String,
    },
}

A representation of the GitHub Event API's event payload.

We only represent the event types that are visible in timelines and haven't been phased out yet.

Variants
CommitComment

A CommitCommentEvent.

Fields

content: String

The comment's body.

commit_id: String

The commit SHA.

id: u64

The comment's ID.

Create

A CreateEvent.

Fields

ref_type: String

The ref's type.

Can be "repository", "branch", or "tag".

ref_name: Option<String>

The ref's name.

None if only a repository was created.

master_branch: String

The master branch of the repository.

repo_description: String

The repository's description (the text right under the tab list).

Delete

A DeleteEvent.

Fields

ref_type: String

The deleted ref's type.

Can be "branch" or "tag".

ref_name: String

The deleted ref's name.

Fork

A ForkEvent.

Fields

new_slug: String

The fork's slug.

Gollum

A GollumEvent (a.k.a. a WikiEvent).

Fields

pages: Vec<GollumPayload>

The affected Wiki pages.

IssueComment

An IssueCommentEvent.

Fields

action: String

The action executed upon a comment.

Can be "created", "edited", or "deleted".

issue: u64

The issue number.

body: String

The issue comment's body text.

id: u64

The issue comment's ID.

Issues

An IssuesEvent.

Fields

action: String

The action executed upon an issue.

Can be "assigned", "unassigned", "labeled", "unlabeled", "opened", "edited", "milestoned", "demilestoned", "closed", or "reopened".

number: u64

The issue number.

title: String

The issue's title.

body: String

The issue body's text.

labels: Vec<String>

The labels the issue has upon it.

Member

A MemberEvent.

Fields

action: String

The action executed upon a user.

Can be only "added".

user: String

The user the action was performed upon.

Public

A PublicEvent.

A repository was made public.

PullRequest

A PullRequestEvent.

Fields

action: String

The action executed upon a PR.

Can be "assigned", "unassigned", "labeled", "unlabeled", "opened", "edited", "closed", or "reopened".

If the action is "closed" and the merged is false, the pull request was closed with unmerged commits. If the action is "closed" and merged is true, the pull request was merged.

number: u64

The PR number.

title: String

The PR's title.

body: String

The PR body's content.

merged: bool

Whether the PR was merged.

If the action is "closed" and the merged is false, the pull request was closed with unmerged commits. If the action is "closed" and merged is true, the pull request was merged.

PullRequestReview

A PullRequestReviewEvent.

Fields

action: String

The action executed upon a PR review.

Can be only "submitted".

pr: u64

The PR number.

state: String

The PR review's state.

Something like "approved".

body: String

The PR review's body.

id: u64

The PR review's ID.

PullRequestReviewComment

A PullRequestReviewCommentEvent.

Fields

action: String

The action executed upon a PR review comment.

Can be "created", "edited", or "deleted".

pr: u64

The PR number.

body: String

The PR review comment's body.

id: u64

The PR review comment's ID.

Push

A PushEvent.

Fields

pushed_ref: String

The full Git ref pushed to.

prev_head: String

The previous HEAD's SHA.

new_head: String

The new HEAD's SHA.

size: u64

The amount of commits pushed.

distinct_size: u64

The amount of distinct commits pushed.

commits: Vec<Commit>

The commits pushed.

Release

A ReleaseEvent.

Fields

action: String

The action executed upon a release.

Can be only "published".

tag_name: String

The released tag's name.

target: String

The released branch's name.

draft: bool

Whether the release is a draft.

prerelease: bool

Whether the release is a prerelease.

name: Option<String>

The release's custom name.

This is None if the release is just a tag and not edited through the GitHub release editor.

body: Option<String>

The release's custom body.

This is None if the release is just a tag and not edited through the GitHub release editor.

Watch

A WatchEvent, or, more aptly, a StarEvent.

Fields

action: String

The action executed upon a repository star.

Can only be "started".

Other

An unhandled event.

Fields

event_type: String

The event type.

Methods

impl EventPayload

fn from(tpe: &str, payload: &JsonValue) -> EventPayload

Parse the payload from the payload key of the event and the type.

Not meant to be used directly, use Event::parse() instead.